Change NumRows to long from int

+++ src/Knet.Kudu.Client/Scanner/ColumnarResultSetParser.cs
@@ -20,7 +20,7 @@ namespace Knet.Kudu.Client.Scanner
 
             _scanSchema = scanSchema;
             _responsePB = scanResponse;
-            NumRows = (int)scanResponse.ColumnarData.NumRows;
+            NumRows = scanResponse.ColumnarData.NumRows;
         }
 
         public override void ParseSidecars(KuduSidecars sidecars)
